test(inflight): move the #4370 marker test out of the save_store graveyard
codex review r2 (#3, #4), independently caught by CI:

- `parent_test_residue` baseline-gate failed: the F1 regression test was added
  inline to `save_store.rs`, whose frozen test-residue ceiling is 730 LoC
  (820 > 730). Raising the ceiling is forbidden (#4269), so migrate the test to
  `save_store/identity_gate.rs`, where the function under test actually lives.
  That file is not a decomposition parent (no sibling dir) and the giant check
  counts production LoC only, so nothing else moves.

- The test resolved the runtime root without pinning it, so an ambient
  `AGENTDESK_ROOT_DIR=~/.adk/release` (present in every release workspace) tripped
  the #3293 live-root assert. Add the standard `shared_test_env_lock` +
  `TestEnvVarGuard` tempdir pin used by its sibling tests.

Verified: passes both with and without an ambient AGENTDESK_ROOT_DIR;
audit --check rc=0; blind-save ratchet 26/26.
