+++ src/services/discord/inflight/save_store/identity_gate.rs
@@ -969,3 +969,125 @@ pub(in crate::services::discord::inflight) fn save_inflight_state_if_matches_ide
         }
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn drain_restart_seed(channel_id: u64, tmux_session_name: &str) -> InflightTurnState {
+        InflightTurnState::new(
+            ProviderKind::Codex,
+            channel_id,
+            Some("adk-test".to_string()),
+            343_742_347_365_974_026,
+            77_010,
+            18,
+            "user prompt".to_string(),
+            Some("session".to_string()),
+            Some(tmux_session_name.to_string()),
+            Some(format!("/tmp/{tmux_session_name}.jsonl")),
+            None,
+            512,
+        )
+    }
+
+    // #4370 F1: the `readopted_from_inflight` marker is a NARROW single-field
+    // patch. It lands on a DrainRestart-preserved row (where the broad
+    // identity-refresh save deliberately refuses `restart_mode` rows), preserving
+    // `restart_mode`; it never resurrects a concurrently-cleared row (`Missing`);
+    // and it refuses to clobber a different turn's row (`IdentityMismatch`).
+    #[test]
+    fn readopted_marker_lands_on_restart_preserved_row_and_never_resurrects() {
+        // #3293: pin the runtime root to a tempdir before any state construction
+        // resolves it, so an ambient `AGENTDESK_ROOT_DIR=~/.adk/release` (every
+        // release workspace has one) cannot make this test touch live state.
+        let _lock = crate::config::shared_test_env_lock()
+            .lock()
+            .unwrap_or_else(|poison| poison.into_inner());
+        let temp = tempfile::TempDir::new().expect("runtime root");
+        let _env_reset = crate::config::TestEnvVarGuard::set_path_after_shared_test_env_lock(
+            "AGENTDESK_ROOT_DIR",
+            temp.path(),
+        );
+        let provider = ProviderKind::Codex;
+
+        // (1) Missing: no durable row → the marker patch does NOT resurrect it.
+        let mut state = drain_restart_seed(44_370, "AgentDesk-codex-4370-drain");
+        let expected = InflightTurnIdentity::from_state(&state);
+        assert_eq!(
+            mark_readopted_from_inflight_if_identity_unchanged_in_root(
+                temp.path(),
+                &provider,
+                state.channel_id,
+                &expected,
+            ),
+            GuardedSaveOutcome::Missing,
+            "an absent row must not be resurrected by the marker patch",
+        );
+
+        // (2) Saved on a DrainRestart-preserved row — where the broad refresh
+        // REFUSES (`restart_mode.is_some()`), proving why F1 needs this narrow
+        // patch instead of `save_inflight_state_if_identity_unchanged`.
+        state.set_restart_mode(InflightRestartMode::DrainRestart);
+        save_inflight_state_in_root(temp.path(), &state).expect("seed restart-preserved row");
+        let expected = InflightTurnIdentity::from_state(&state);
+        assert_eq!(
+            save_inflight_state_if_identity_unchanged_in_root(
+                temp.path(),
+                &state,
+                "test::readopted_marker_broad_refresh_refuses",
+            ),
+            GuardedSaveOutcome::IdentityMismatch,
+            "the broad identity-refresh save must keep refusing restart_mode rows",
+        );
+        assert_eq!(
+            mark_readopted_from_inflight_if_identity_unchanged_in_root(
+                temp.path(),
+                &provider,
+                state.channel_id,
+                &expected,
+            ),
+            GuardedSaveOutcome::Saved,
+        );
+
+        let persisted_path = inflight_state_path(temp.path(), &provider, state.channel_id);
+        let persisted: InflightTurnState = serde_json::from_str(
+            &std::fs::read_to_string(&persisted_path).expect("read persisted inflight"),
+        )
+        .expect("parse persisted inflight");
+        assert!(
+            persisted.readopted_from_inflight,
+            "the marker must land on the restart-preserved row"
+        );
+        assert_eq!(
+            persisted.restart_mode,
+            Some(InflightRestartMode::DrainRestart),
+            "restart_mode must be preserved by the single-field patch"
+        );
+
+        // (3) Idempotent: a re-mark of an already-marked row is a `Saved` no-op.
+        assert_eq!(
+            mark_readopted_from_inflight_if_identity_unchanged_in_root(
+                temp.path(),
+                &provider,
+                state.channel_id,
+                &expected,
+            ),
+            GuardedSaveOutcome::Saved,
+        );
+
+        // (4) IdentityMismatch: a different turn identity must not be clobbered.
+        let mut other = state.clone();
+        other.user_msg_id = 99_999;
+        let mismatched = InflightTurnIdentity::from_state(&other);
+        assert_eq!(
+            mark_readopted_from_inflight_if_identity_unchanged_in_root(
+                temp.path(),
+                &provider,
+                state.channel_id,
+                &mismatched,
+            ),
+            GuardedSaveOutcome::IdentityMismatch,
+        );
+    }
+}
